// src/services/category/enum.ts
export enum ECategoryType {
  MORTGAGE = "Mortgage",
  FOOD = "food",
  UTILITIES = "utilities",
  BILLS = "bills",
  SHOPPING = "shopping",
  TRANSPORTATION = "transportation",
  INSURANCE = "insurance",
  HEALTHCARE = "healthcare",
  CLOTHING = "clothing",
  OTHERS = "others",
}

export enum ECategoryTypeColor {
  MORTGAGE = "#d1d5db",
  FOOD = "#10b981",
  UTILITIES = "#f5a316",
  BILLS = "#f43f5e",
  SHOPPING = "#0ea5e9",
  TRANSPORTATION = "#38bdf8",
  INSURANCE = "#ec4899",
  HEALTHCARE = "#6366f1",
  CLOTHING = "#1e293b",
  OTHERS = "#111",
}